docs: Add comments to develop better understanding of Chapter 7 project
Add comments to get a better understanding of the the problem and the solution at a conceptual level by understanding the behavior of the code that has been used to implement the abstract ideas of the author.

Understanding the code of project given in Chapter 7 is key to solving the programming exercises of the said chapter.

The code, while simple enough to be understood by beginners, deals with a lot of high level concepts & abstractions which adds a layer of complexity.

Thus, the author's solution to the key problem (final objective of the project) has the potential to confuse the learners if they try to rush through this chapter (7) or has failed to properly understand the concepts taught in the earlier chapters (1-6).