docs: Modify comments to correct mistakes in my understanding of code
After further inspection of the code and re-reading of  the chapter, I came to realize that I have misunderstood certain key ideas behind the code.

Due to which, there are certain comments which reflect my misunderstood conclusions. Hence, I have modified the problematic sections of comments to ensure that the reader is not misguided and can form the correct mental models required for complete understanding of the code.