docs: add: A simulation function with implementation of 3 robot types
The code (along with its associated documentation) can be split into four broad categories:

1. A simulator function which will simulate the process of a type of robot picking up & delivering the parcels in a virtual village of Meadowfield.

2. The implementation (with proper documentation) & simulated run of randomRobot which uses a naive approach to deliver parcels. It uses the most inefficient approach.

3. The implementation (along with associated documentation) and the simulated run of routeRobot which uses an approach similar to that of real world mail delivery systems. 

While a lot more efficient, it relies on a hard coded route (which passes through every place in the village) to complete its job which usually means that it fails to find the most effective path to deliver parcels.

4. The implementation and the simulated run of goalOrientedRobot, without any kind of associated documentation, to aid the understanding of the reader. 

This approach tries to actively find the shortest path that the robot could take to deliver all the parcels. It is, more often than not, finds the quickest way to solve the problem.