docs: Add documentation to understand behavior of Goal Oriented Robot
With the addition of the comments to the functions associated with the goal oriented robot (robot type which tries to actively find the shortest route between places), the line by line inspection and description of Chapter 7's project code is now complete.

I have also added additional console print statements to highlight the approach of the aforementioned robot type, which will further aid in a more complete understanding of the code.

I have also added comments to the findRoute utility function which is used in the implementation of the goal oriented robot.