add: Line by line analysis of the Chapter 12's project code
I have added comments to the code given in the Chapter 12's project where the author has created a super simple language called Egg...

The programming language consists of a parser, a basic interpreter, scopes, language constructs (if, while, etc) and functions.

The code is quite simple but the design of the language and how each individual component interacts with one another, could cause confusion to readers who try to rush this chapter...

Also, an understanding of the fundamental ideas behind Egg language & its syntax is important if one wishes to decipher how the code works...

Hence, I have decided to add a line-by-line analysis to help readers demystify the code of the project. Understanding it will be key to solving the programming exercises related to this chapter.