modify: 'Tabs' solution to change style of currently selected tabs
I have modified the solution of 'Tabs' solution to switch the style of currently selected tab (as mentioned in the requirements) to highlight which tab is selected.

Also, this would mean that I should reset the style of previously selected tab to its default state.

I have decided to add a <style> tag to the HTML document to keep the separation between presentation details of the HTML elements and the logical implementation of the script.

To achieve this, I have added a `selected` CSS class and then, using JavaScript, manipulated the DOM to add/remove it when required.

Also, I have refactored the code to display the contents of the tabs by adding a `hidden` CSS class to the `div` tags in the `tab-panel` tag and then, toggling it on/off as per the user's wishes.